fixed bug in imagetatgging management command, related to new groundtruth tags

+++ imagetagging/management/commands/populate_imagetagging_db.py
@@ -18,16 +18,24 @@ class Command(BaseCommand):
     def add_arguments(self, parser):
         parser.add_argument('csv_file', nargs='+', type=str)
 
+        # Named (optional) arguments
+        parser.add_argument(
+            '--update',
+            action='store_true',
+            help='clear groundtruth tags before adding new',
+        )
+
     def handle(self, *args, **options):
         self.stdout.write("importing.. \n")
 
-        GroundTruthTag.objects.all().delete()
-        ImageTask.objects.all().delete()
+        if options['update']:
+            GroundTruthTag.objects.all().delete()
+        # ImageTask.objects.all().delete()
 
         # if User.objects.filter(is_superuser=True).exists():
         #     superuser = User.objects.filter(is_superuser=True)[0]
         # else:
-        superuser = User.objects.create_superuser('groundtruth', email='', password='password')
+        # superuser = User.objects.create_superuser('groundtruth', email='', password='password')
 
         for filename in options['csv_file']:
             print('filename:', filename)
@@ -40,10 +48,14 @@ class Command(BaseCommand):
                 tags = line[1:]
 
                 filename = '%d.jpg' % image_number
+                filename = os.path.join('imagetagging', 'images', filename)
                 # create ImageTask
-                image_task = ImageTask(next_task = None)
-                image_task.image.name = os.path.join('imagetagging', 'images', filename)
-                image_task.save()
+                if options['update']:
+                    image_task = ImageTask.objects.get(image=filename)
+                else:
+                    image_task = ImageTask(next_task=None)
+                    image_task.image.name = filename
+                    image_task.save()
 
                 # create tags
                 for current_tag in tags:
@@ -52,12 +64,13 @@ class Command(BaseCommand):
                         image_task=image_task
                     )
 
-            # link images
-            all_tasks = ImageTask.objects.all()
-            prev_image_task = all_tasks.last()
-            for image_task in all_tasks:
-                prev_image_task.next_task = image_task
-                prev_image_task.save()
-                prev_image_task = image_task
+            if not options['update']:
+                # link images
+                all_tasks = ImageTask.objects.all()
+                prev_image_task = all_tasks.last()
+                for image_task in all_tasks:
+                    prev_image_task.next_task = image_task
+                    prev_image_task.save()
+                    prev_image_task = image_task
 
         self.stdout.write("\n..done\n")
